merge new version

+++ src/prod/src/data/interop/ReliableCollectionServiceStandalone/ReliableCollectionServiceStandalone.cpp
@@ -6,8 +6,28 @@
 #include "stdafx.h"
 #include <stdlib.h>
 
+#include <ktlevents.um.h>
+
 using namespace Common;
 using namespace Data::Integration;
+
+enum ReliableCollectionMode : uint64_t
+{
+    /// <summary>
+    /// Recover from previous checkpoint, if available
+    /// </summary>
+    OpenOrCreate = 0x4,
+
+    /// <summary>
+    /// No Recovery, removes previous checkpoints and provides fresh reliable collections.
+    /// </summary>
+    CreateAsNew = 0x8,
+
+    /// <summary>
+    /// Use Volatile Reliable collections
+    /// </summary>
+    Volatile = 0x10
+};
 
 namespace Data
 {
@@ -17,6 +37,28 @@ namespace Data
         using namespace ktl;
         using namespace Data::Utilities;
         using namespace Data::TStore;
+        using namespace Data::Log;
+
+        class InitConfig
+        {
+        public:
+            InitConfig() {
+#if defined(PLATFORM_UNIX)
+        #define DEFINE_STRING_RESOURCE(id,str) StringResourceRepo::Get_Singleton()->AddResource(id, str);
+            #include "rcstringtbl.inc"
+#endif
+            }
+        };
+
+        InitConfig __init;
+
+        class RCStandaloneCommonConfig : ComponentConfig
+        {
+            DECLARE_SINGLETON_COMPONENT_CONFIG(RCStandaloneCommonConfig, "RCStandaloneCommonConfig")
+
+            PUBLIC_CONFIG_ENTRY(int, L"Trace/Etw", Sampling, 1, ConfigEntryUpgradePolicy::Dynamic);
+            PUBLIC_CONFIG_ENTRY(int, L"Trace/Etw", Level, 4, ConfigEntryUpgradePolicy::Dynamic);
+        };
 
         class CExportTests
         {
@@ -24,10 +66,9 @@ namespace Data
             CExportTests(wstring const & testName)
             {
                 testName_ = testName;
-                epoch_ = 0;
             }
 
-            void Setup();
+            void Setup(__in ReliableCollectionMode mode);
             void Cleanup();
             void OpenReplica();
             void RequestCheckpointAfterNextTransaction();
@@ -63,7 +104,7 @@ namespace Data
             void CloseReplica();
             Awaitable<void> OpenReplicaAsync();
 
-            CommonConfig config; // load the config object as its needed for the tracing to work
+            RCStandaloneCommonConfig config; // load the config object as its needed for the tracing to work
             KtlSystem * underlyingSystem_;
 
             KGuid pId_;
@@ -76,35 +117,48 @@ namespace Data
             wstring testFolderPath_;
             wstring testName_;
             LONGLONG epoch_;
+            BOOLEAN hasPersistedState_;
+
+            static const wstring DEFAULT_TEST_PARTITION_GUID;
+            static const long DEFAULT_TEST_REPLICA_ID;
         };
 
-        void CExportTests::Setup()
+        const wstring CExportTests::DEFAULT_TEST_PARTITION_GUID = L"45AEFD6F-1CCE-4A2E-8C58-9AD7D6A9CC7C";
+        const long CExportTests::DEFAULT_TEST_REPLICA_ID = 1234567890;
+
+        void CExportTests::Setup(__in ReliableCollectionMode mode)
         {
             NTSTATUS status;
 
             testFolderPath_ = CreateFileName(testName_);
             wstring TestLogFileName = testName_.append(L".log");
 
-            Directory::Delete_WithRetry(testFolderPath_, true, true);
-
             status = KtlSystem::Initialize(FALSE, &underlyingSystem_); 
             ASSERT_IFNOT(NT_SUCCESS(status), "Status not success: {0}", status); 
 
-            underlyingSystem_->SetStrictAllocationChecks(TRUE); 
-            KAllocator & allocator = underlyingSystem_->PagedAllocator(); 
-            int seed = GetTickCount(); 
-            Common::Random r(seed); 
-            rId_ = r.Next(); 
-            pId_.CreateNew(); 
-            prId_ = PartitionedReplicaId::Create(pId_, rId_, allocator); 
+            bool allowRecovery = !(mode & ReliableCollectionMode::CreateAsNew);
+            hasPersistedState_ = !(mode & ReliableCollectionMode::Volatile);
+            if (!allowRecovery)
+               Directory::Delete_WithRetry(testFolderPath_, true, true);
+
+            rId_ = DEFAULT_TEST_REPLICA_ID;
+            Guid guid(DEFAULT_TEST_PARTITION_GUID);
+            KGuid paritionGuid(guid.AsGUID());
+            pId_ = paritionGuid;
+            underlyingSystem_->SetStrictAllocationChecks(TRUE);
+            KAllocator & allocator = underlyingSystem_->PagedAllocator();
+            prId_ = PartitionedReplicaId::Create(pId_, rId_, allocator);
 
             KtlLogger::SharedLogSettingsSPtr sharedLogSettings;
             InitializeKtlConfig(testFolderPath_, TestLogFileName, underlyingSystem_->PagedAllocator(), sharedLogSettings);
 
+            TraceProvider::GetSingleton()->SetDefaultLevel(TraceSinkType::ETW, LogLevel::Noise);
+            RegisterKtlTraceCallback(Common::ServiceFabricKtlTraceCallback);
+
             status = Data::Log::LogManager::Create(underlyingSystem_->PagedAllocator(), logManager_);
             CODING_ERROR_ASSERT(NT_SUCCESS(status));
 
-            status = SyncAwait(logManager_->OpenAsync(CancellationToken::None, sharedLogSettings));
+            status = SyncAwait(logManager_->OpenAsync(CancellationToken::None, sharedLogSettings, KtlLoggerMode::InProc));
             CODING_ERROR_ASSERT(NT_SUCCESS(status));
 
             replica_ = CreateReplica();
@@ -117,14 +171,17 @@ namespace Data
             wstring testFolder = Path::Combine(testFolderPath_, L"work");
             StateProviderFactory::SPtr factory;
             StateProviderFactory::Create(underlyingSystem_->PagedAllocator(), factory);
-            
+
             Replica::SPtr replica = Replica::Create(
                 pId_,
                 rId_,
                 testFolder,
                 *logManager_,
                 underlyingSystem_->PagedAllocator(),
-                factory.RawPtr());
+                factory.RawPtr(),
+                nullptr,
+                nullptr,
+                hasPersistedState_);
             return replica;
         }
 
@@ -137,9 +194,13 @@ namespace Data
         {
             co_await replica_->OpenAsync();
 
-            epoch_++;
-            FABRIC_EPOCH epoch1; epoch1.DataLossNumber = epoch_; epoch1.ConfigurationNumber = epoch_; epoch1.Reserved = nullptr;
-            co_await replica_->ChangeRoleAsync(epoch1, FABRIC_REPLICA_ROLE_PRIMARY);
+            FABRIC_EPOCH epoch;
+            NTSTATUS status = replica_->TxnReplicator->GetCurrentEpoch(epoch);
+            CODING_ERROR_ASSERT(NT_SUCCESS(status));
+            LONG64 currentConfigurationNumber = epoch.ConfigurationNumber;
+            currentConfigurationNumber++;
+            epoch.ConfigurationNumber = currentConfigurationNumber;
+            co_await replica_->ChangeRoleAsync(epoch, FABRIC_REPLICA_ROLE_PRIMARY);
 
             replica_->SetReadStatus(FABRIC_SERVICE_PARTITION_ACCESS_STATUS_GRANTED);
             replica_->SetWriteStatus(FABRIC_SERVICE_PARTITION_ACCESS_STATUS_GRANTED);
@@ -192,7 +253,10 @@ namespace Data
         wstring CExportTests::CreateFileName(
             __in wstring const & folderName)
         {
-            wstring testFolderPath = Directory::GetCurrentDirectoryW();
+            wstring testFolderPath = L"";
+            if (!Environment::GetEnvironmentVariable(L"Fabric_RCStandaloneWorkDirOverride", testFolderPath, NOTHROW()))
+                testFolderPath = Directory::GetCurrentDirectoryW();
+
             Path::CombineInPlace(testFolderPath, folderName);
 
             return testFolderPath;
@@ -217,29 +281,33 @@ namespace Data
             KInvariant(sharedLogFileName->LengthInBytes() + sizeof(WCHAR) < 512 * sizeof(WCHAR)); // check to make sure there is space for the null terminator
             KMemCpySafe(&settings->Path[0], 512 * sizeof(WCHAR), sharedLogFileName->operator PVOID(), sharedLogFileName->LengthInBytes());
             settings->Path[sharedLogFileName->LengthInBytes() / sizeof(WCHAR)] = L'\0'; // set the null terminator
-            settings->LogContainerId.GetReference().CreateNew();
+            settings->LogContainerId = KGuid(KtlLogger::Constants::DefaultApplicationSharedLogId.AsGUID()); // always use staging log for standalone
             settings->LogSize = 1024 * 1024 * 512; // 512 MB.
             settings->MaximumNumberStreams = 0;
             settings->MaximumRecordSize = 0;
-            settings->Flags = static_cast<ULONG>(Data::Log::LogCreationFlags::UseSparseFile);
+            settings->Flags = static_cast<ULONG>(Data::Log::LogCreationFlags::UseNonSparseFile);
             sharedLogSettings = make_shared<KtlLogger::SharedLogSettings>(std::move(settings));
         };
+
     }
 }
 
-static Data::Interop::CExportTests* t = nullptr;
+static Data::Interop::CExportTests* tests = nullptr;
 
-extern "C" HRESULT Initialize(
-    __in LPCWSTR testname, 
-    __out void** txnReplicator, 
+extern "C" HRESULT ReliableCollectionStandalone_InitializeEx(
+    __in LPCWSTR testname,
+    __in ReliableCollectionMode mode,
+    __out void** txnReplicator,
     __out void** partition,
     __out GUID* partitionId,
     __out int64_t* replicaId)
 {
     wstring env;
-    HRESULT hr = ReliableCollectionRuntime_Initialize(RELIABLECOLLECTION_API_VERSION);
-    if (!SUCCEEDED(hr))
-        return hr;
+
+    //
+    // Enable KTL tracing
+    //
+    EventRegisterMicrosoft_Windows_KTL();
 
     if (Environment::GetEnvironmentVariable(L"SF_RELIABLECOLLECTION_MOCK", env, NOTHROW()))
     {
@@ -248,39 +316,56 @@ extern "C" HRESULT Initialize(
         return S_OK;
     }
 
-    t = new Data::Interop::CExportTests(testname);
-    t->Setup();
-    *txnReplicator = t->GetTxnReplicator();
+    tests = new Data::Interop::CExportTests(testname);
+    tests->Setup(mode);
+    *txnReplicator = tests->GetTxnReplicator();
     *partition = nullptr;
-    *partitionId = t->PartitionId;
-    *replicaId = t->ReplicaId;
+    *partitionId = tests->PartitionId;
+    *replicaId = tests->ReplicaId;
     return S_OK;
 }
 
-extern "C" void ResetReplica(__out void** txnReplicator)
+extern "C" HRESULT ReliableCollectionStandalone_Initialize(
+    __in LPCWSTR testname,
+    __in bool allowRecovery,
+    __out void** txnReplicator,
+    __out void** partition,
+    __out GUID* partitionId,
+    __out int64_t* replicaId)
 {
-    t->ResetReplica();
-    *txnReplicator = t->GetTxnReplicator();
+    ReliableCollectionMode reliableCollectionMode = allowRecovery ? ReliableCollectionMode::OpenOrCreate : ReliableCollectionMode::CreateAsNew;
+    return ReliableCollectionStandalone_InitializeEx(testname, reliableCollectionMode, txnReplicator, partition, partitionId, replicaId);
 }
 
-extern "C" void OpenReplica()
+extern "C" void ReliableCollectionStandalone_ResetReplica(__out void** txnReplicator)
 {
-    t->OpenReplica();
+    tests->ResetReplica();
+    *txnReplicator = tests->GetTxnReplicator();
 }
 
-extern "C" void RequestCheckpointAfterNextTransaction()
+extern "C" void ReliableCollectionStandalone_OpenReplica()
 {
-    t->RequestCheckpointAfterNextTransaction();
+    tests->OpenReplica();
 }
 
-extern "C" void Cleanup()
+extern "C" void ReliableCollectionStandalone_RequestCheckpointAfterNextTransaction()
 {
-    if (t != nullptr)
+    tests->RequestCheckpointAfterNextTransaction();
+}
+
+extern "C" void ReliableCollectionStandalone_Cleanup()
+{
+    if (tests != nullptr)
     {
-        t->Cleanup();
-        delete t;
-        t = nullptr;
+        tests->Cleanup();
+        delete tests;
+        tests = nullptr;
     }
+
+    //
+    // Disable KTL Tracing
+    //
+    EventUnregisterMicrosoft_Windows_KTL();
 }
 
 extern "C" HRESULT FabricGetReliableCollectionApiTable(uint16_t apiVersion, ReliableCollectionApis* reliableCollectionApis)
@@ -301,6 +386,8 @@ extern "C" HRESULT FabricGetReliableCollectionApiTable(uint16_t apiVersion, Reli
     return S_OK;
 }
 
+// Dummy methods required to satisfy link of ReliableCollectionServiceStandalone.dll
+// These methods are not used at runtime for ReliableCollectionServiceStandalone
 HRESULT FabricGetActivationContext( 
     REFIID riid,
     void **activationContext)
@@ -320,6 +407,19 @@ HRESULT FabricLoadReplicatorSettings(
     UNREFERENCED_PARAMETER(configurationPackageName);
     UNREFERENCED_PARAMETER(sectionName);
     UNREFERENCED_PARAMETER(replicatorSettings);
+    return S_OK;
+}
+
+HRESULT FabricLoadSecurityCredentials(
+    IFabricCodePackageActivationContext const * codePackageActivationContext,
+    LPCWSTR configurationPackageName,
+    LPCWSTR sectionName,
+    IFabricSecurityCredentialsResult ** securitySettings)
+{
+    UNREFERENCED_PARAMETER(codePackageActivationContext);
+    UNREFERENCED_PARAMETER(configurationPackageName);
+    UNREFERENCED_PARAMETER(sectionName);
+    UNREFERENCED_PARAMETER(securitySettings);
     return S_OK;
 }
